+++ packages/nimble-blazor/Examples/Demo.Hybrid/App.xaml.cs
@@ -1,0 +1,11 @@
+﻿namespace Demo.Hybrid
+{
+    using System.Windows;
+
+    /// <summary>
+    /// Interaction logic for App.xaml.
+    /// </summary>
+    public partial class App : Application
+    {
+    }
+}