+++ packages/nimble-components/src/table-column/text/group-header-view/index.ts
@@ -1,9 +1,9 @@
 import { DesignSystem } from '@microsoft/fast-foundation';
-import type { TableColumnTextColumnConfig } from '..';
 import type { TableStringFieldValue } from '../../../table/types';
 import { TableColumnTextGroupHeaderViewBase } from '../../text-base/group-header-view';
 import { template } from '../../text-base/group-header-view/template';
 import { styles } from '../../text-base/group-header-view/styles';
+import type { TableColumnWithPlaceholderColumnConfig } from '../../base/types';
 
 declare global {
     interface HTMLElementTagNameMap {
@@ -16,7 +16,7 @@ declare global {
  */
 export class TableColumnTextGroupHeaderView extends TableColumnTextGroupHeaderViewBase<
 TableStringFieldValue,
-TableColumnTextColumnConfig
+TableColumnWithPlaceholderColumnConfig
 > {
     public override get text(): string {
         return this.groupHeaderValue!;