+++ packages/nimble-components/src/table/components/header/template.ts
@@ -7,7 +7,11 @@ import { TableColumnSortDirection } from '../../types';
 
 // prettier-ignore
 export const template = html<TableHeader>`
-    <template role="columnheader" aria-sort="${x => x.ariaSort}">
+    <template role="columnheader"
+        aria-sort="${x => x.ariaSort}"
+        ${'' /* Prevent header double clicks from selecting text */}
+        @mousedown="${(_x, c) => !((c.event as MouseEvent).detail > 1)}"
+    >
         <slot></slot>
 
         ${when(x => x.sortDirection === TableColumnSortDirection.ascending, html`